Adding Package_Services functions

+++ Package_Services/src/main/java/cn/lingnan/dto/PackageBoxInformation.java
@@ -7,9 +7,9 @@ import lombok.Data;
 public class PackageBoxInformation {
     private Object pbiId;
 
-    private Integer packageId;
+    private Long packageId;
 
-    private Integer lockerId;
+    private Long lockerId;
 
     private Object lockerBoxId;
